feat: add editing cover id optional param to the capacity for product route
